Split manager and app_init to satisfy the source-size policy
Both files crossed the 1000-line hard limit as this branch grew.

The identity accessors leaving manager.rs are a cohesive unit rather than
an arbitrary cut: they are the runtime layer's answer to `which process?`,
and the point of this issue is that the question has three distinct answers
that must never substitute for one another. Grouping them makes that
explicit.

app_init.rs keeps its tests, now included via #[path] exactly as the
runtime manager's tests already are. All 15 move across unchanged.

Refs #543
